+++ lib/src/inspection/camera_summary.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-
 import '../const.dart';
 import '../theme.dart';
 import '../widgets/buttons.dart';