+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 0.0.4
+
+# Change Log
+
+## 0.5.0 - 2022-08-04
+### Changes
+* Added inspection to purchase process for both platforms
+* Added payment method to the purchase process
+
+
+## 0.0.4 -  2022-08-04
 
 * Initial beta release.